Created data class in datset.py moved preprocessing methods from module.py to dataset.py
